+ причесал мобакт и выделил команду и обработку трека.
