+++ src/graph.h
@@ -1,0 +1,9 @@
+#ifndef BYLINS_GRAPH_H
+#define BYLINS_GRAPH_H
+
+#include "chars/char.hpp"
+
+int find_first_step(room_rnum src, room_rnum target, CHAR_DATA * ch);
+
+
+#endif //BYLINS_GRAPH_H
