+++ src/interpreter.cpp
@@ -17,6 +17,8 @@
 
 #include "cmd/retreat.h"
 #include "cmd/order.h"
+#include "cmd/track.h"
+
 #include "skills/manadrain.h"
 #include "skills/flee.h"
 #include "skills/bash.h"
@@ -227,7 +229,6 @@ void do_horseoff(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_horseput(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_horseget(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_horsetake(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
-void do_hidetrack(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_hidemove(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_fit(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_force(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
@@ -315,7 +316,6 @@ void do_teleport(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_tell(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_time(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_toggle(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
-void do_track(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_sense(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_unban(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_ungroup(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
